src/requirements.txt
beautifulsoup4==4.11.2
google-api-python-client==2.83.0
google-cloud-compute==1.11.0
pandas==1.5.3
paramiko==3.1.0
selenium==4.8.2
